compiles ; norminette pleasing for export_builtin ; with all the files that time
