+++ utils/is_chr_in_str.c
@@ -1,0 +1,26 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   is_chr_in_str.c                                    :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: nstoutze <nstoutze@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/11/15 09:57:53 by nstoutze          #+#    #+#             */
+/*   Updated: 2023/11/15 09:58:08 by nstoutze         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+int	is_chr_in_str(char *s, char c)
+{
+	int	i;
+	
+	i = -1;
+	while (s[++i])
+	{
+		if (s[i] == c)
+			return (i);
+	}
+	return (0);
+}
